refactor(web): extract shared ChipInput core
acl and forward duplicated the entire chip-input core (handler logic,
the flush handle, and the per-chip markup). Extract a useChipInput hook,
a Chip component, and the shared types into components/chip-input. Each
page keeps its own wrapper: acl its bulk-paste/popover/filter toolbar,
forward its flat input, and their distinct wildcard chips.
