+++ web/src/components/chip-input/index.ts
@@ -1,0 +1,3 @@
+export { useChipInput } from './useChipInput';
+export { Chip } from './Chip';
+export type { ChipKind, ChipInputProps, ChipInputHandle } from './types';
